Replaced FailureWidget with FailureBanner
